v2

+++ src/sensors/light_sensor.h
@@ -1,7 +1,7 @@
 #ifndef LIGHT_SENSOR_H
 #define LIGHT_SENSOR_H
-
+#include <stdint.h>
 void light_sensor_init(void);
-int light_sensor_read(void);
+uint16_t light_sensor_read(void);
 
 #endif
